feat: implementação recapctha

+++ arena-lagoa-beach/components/auth/LoginForm.tsx
@@ -4,7 +4,8 @@ import Link from "next/link";
 import { useState } from "react";
 import { useRouter } from "next/navigation";
 import { login } from "@/app/services/authLogin";
-import PopupModelo from "@/components/ui/PopupModelo"; // ajusta o caminho
+import PopupModelo from "@/components/ui/PopupModelo";
+import Recaptcha from "@/components/recaptcha/recaptcha";
 
 export function LoginForm() {
   const router = useRouter();
@@ -14,7 +15,10 @@ export function LoginForm() {
   const [lembrar, setLembrar] = useState(false);
   const [loading, setLoading] = useState(false);
 
-  // 🔥 estados do popup
+  // 🔐 captcha
+  const [captchaToken, setCaptchaToken] = useState<string | null>(null);
+
+  // 🔥 popup
   const [modalOpen, setModalOpen] = useState(false);
   const [modalMessage, setModalMessage] = useState("");
 
@@ -27,15 +31,21 @@ export function LoginForm() {
       return;
     }
 
+    // ✅ valida captcha
+    if (!captchaToken) {
+      setModalMessage("Confirme que você não é um robô 🤖");
+      setModalOpen(true);
+      return;
+    }
+
     try {
       setLoading(true);
 
-      const res = await login({ email, senha });
+      const res = await login({ email, senha, captchaToken });
 
       // salva token
       localStorage.setItem("token", res.token);
 
-      // ✅ sucesso → NÃO mostra popup
       router.push("/home");
     } catch (err: any) {
       setModalMessage(err.response?.data?.message || "Erro ao fazer login");
@@ -82,12 +92,19 @@ export function LoginForm() {
               onChange={(e) => setLembrar(e.target.checked)}
               className="h-4 w-4 accent-[#C2E96A]"
             />
-            <label className="text-sm text-gray-700">Lembre-se de mim</label>
+            <label className="text-sm text-gray-700">
+              Lembre-se de mim
+            </label>
           </div>
 
           <span className="text-sm text-red-500 hover:text-red-700 cursor-pointer">
             Esqueci minha senha
           </span>
+        </div>
+
+        {/* RECAPTCHA */}
+        <div className="flex justify-center mb-4">
+          <Recaptcha onChange={setCaptchaToken} />
         </div>
 
         <div className="flex justify-center">
@@ -108,7 +125,7 @@ export function LoginForm() {
         </span>
       </form>
 
-      {/* 🔥 POPUP DE ERRO */}
+      {/* 🔥 POPUP */}
       <PopupModelo
         isOpen={modalOpen}
         onClose={() => setModalOpen(false)}
